+++ src/shader_program.cpp
@@ -1,9 +1,9 @@
 /*
- *    Due Date: 2020-11-15
+ *    Due Date: 2020-12-13
  *      Author: Tim Alberts (timothy.alberts@snhu.edu)
  *       Class: CS-499-X2150 Computer Science Capstone 20EW2
  *  Instructor: Joseph Conlan
- *  	Module: 3-2
+ *  	Module: 7-1
  */
 
 #include "shader_program.h"
@@ -16,7 +16,7 @@ ShaderProgram::~ShaderProgram() {
 	glDeleteProgram(program_handle_);
 }
 
-bool ShaderProgram::LoadShadersFiles(const char* vsFilename, const char* fsFilename) {
+bool ShaderProgram::LoadShadersFiles(const std::string& vsFilename, const std::string& fsFilename) {
 	std::string vsString = fileToString(vsFilename);
 	std::string fsString = fileToString(fsFilename);
 	return LoadShaders(vsString.c_str(), fsString.c_str());