+++ src/shader_program.h
@@ -1,9 +1,9 @@
 /*
- *    Due Date: 2020-11-15
+ *    Due Date: 2020-12-13
  *      Author: Tim Alberts (timothy.alberts@snhu.edu)
  *       Class: CS-499-X2150 Computer Science Capstone 20EW2
  *  Instructor: Joseph Conlan
- *  	Module: 3-2
+ *  	Module: 7-1
  */
 
 #ifndef SHADER_PROGRAM_H_
@@ -29,7 +29,7 @@ class ShaderProgram {
 		kProgram
 	};
 
-	bool LoadShadersFiles(const char* vsFilename, const char* fsFilename);
+	bool LoadShadersFiles(const std::string& vsFilename, const std::string& fsFilename);
 	bool LoadShaders(const GLchar* vsSourcePtr, const GLchar* fsSourcePtr);
 	void Use();